update validate the user updating or deleting a contact is the user who created the contact.

+++ controllers/contactController.js
@@ -13,8 +13,9 @@ const Contact = require("../models/contactModel");
 //@desc Get all contacts
 //@route GET /api/contacts
 //@access private
+// the user that we found with .find() will be getting all the contacts they created
 const getContacts = asyncHandler(async (req, res) => {
-  const contacts = await Contact.find();
+  const contacts = await Contact.find({ user_id: req.user.id });
   res.status(200).json(contacts);
 });
 
@@ -33,6 +34,7 @@ const createContact = asyncHandler(async (req, res) => {
     name,
     email,
     phone,
+    user_id: req.user.id
   });
 
   res.status(201).json(contact);
@@ -60,6 +62,13 @@ const updateContact = asyncHandler(async (req, res) => {
     throw new Error("Contact not found");
   }
 
+  // this is to check for the user who created the contact is
+  // the current user who is updating it
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("User don't have permission to update other user contacts");
+  }
+
   const updatedContact = await Contact.findByIdAndUpdate(
     req.params.id,
     req.body,
@@ -78,7 +87,15 @@ const deleteContact = asyncHandler(async (req, res) => {
     throw new Error("Contact not found");
   }
 
+  // this is to check for the user who created the contact is
+  // the current user who is deleting it
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("User don't have permission to update other user contacts");
+  }
+
   const deletedContact = await Contact.findByIdAndDelete(req.params.id);
+  // await Contact.deleteOne({ _id: req.params.id }); 
 
   res.status(200).json(deletedContact);
 });
